fixing auth for taboola
