+++ packages/destination-actions/src/destinations/taboola-actions/syncAudience/index.ts
@@ -107,7 +107,6 @@ const action: ActionDefinition<Settings, Payload, AudienceSettings> = {
     }
   },
   perform: (request, { payload, audienceSettings }) => {
-
     if (!payload.external_audience_id) {
       throw new IntegrationError('Bad Request: payload.external_audience_id missing.', 'INVALID_REQUEST_DATA', 400)
     }
@@ -120,27 +119,10 @@ const action: ActionDefinition<Settings, Payload, AudienceSettings> = {
       )
     }
 
-    if (!audienceSettings) {
-      throw new IntegrationError('Bad Request: no audienceSettings found.', 'INVALID_REQUEST_DATA', 400)
-    }
-
-    if (!audienceSettings.account_id) {
-      throw new IntegrationError('Bad Request: no audienceSettings.account_id found.', 'INVALID_REQUEST_DATA', 400)
-    }
-
     const taboolaClient = new TaboolaClient(request, [payload], audienceSettings)
     return taboolaClient.sendToTaboola()
   },
-  performBatch: async (request, { payload: payloads, audienceSettings }) => {
-
-    if (!audienceSettings) {
-      throw new IntegrationError('Bad Request: no audienceSettings found.', 'INVALID_REQUEST_DATA', 400)
-    }
-
-    if (!audienceSettings.account_id) {
-      throw new IntegrationError('Bad Request: no audienceSettings.account_id found.', 'INVALID_REQUEST_DATA', 400)
-    }
-
+  performBatch: (request, { payload: payloads, audienceSettings }) => {
     const taboolaClient = new TaboolaClient(request, payloads, audienceSettings)
     return taboolaClient.sendToTaboola()
   }
